requirements.txt
Django==3.1.4
django-bootstrap4==2.3.1
django-ckeditor==6.0.0
django-crispy==1.8.1
django-taggit==1.3.0
djangorestframework==3.12.2
psycopg2-binary==2.8.6
